refactor(): move sideNav to helpers
This took way longer than it should. Gotta love typos that somehow
don't throw errors! (Wishes we were using TypeScript!) :wink2:
