+++ src/helpers/index.js
@@ -11,6 +11,7 @@ export {landing} from './landing'
 export {appMenu} from './appMenu'
 export {mobileFrame, desktopFrame} from './frame'
 export {projectForm} from './projectForm'
+export {sideNav} from './sideNav'
 
 // helper function to attach react components to the snabbdom
 // some need to be attached on 'update', others on 'insert', not sure why
